+++ src/Fields/Types/FileField.php
@@ -43,11 +43,19 @@ class FileField extends MediaField implements Field
 
         /** @var Asset $asset */
         foreach ($assets as $asset) {
+
+            $thumbUrl = $asset->url('thumb');
+
+            // If the conversions haven't run yet, we'll use the original image until they are uploaded
+            if(!file_exists(public_path($thumbUrl))) {
+                $thumbUrl = $asset->url();
+            }
+
             $files[] = (object)[
                 'id'       => $asset->id,
                 'filename' => $asset->filename(),
                 'url'      => $asset->url(),
-                'thumbUrl' => $asset->url('thumb'),
+                'thumbUrl' => $thumbUrl,
                 'mimetype' => $asset->getMimeType(),
                 'isImage'  => ($asset->getExtensionType() == 'image'),
                 'size'     => $asset->getSize(),