+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All Notable changes to the `chief` application template will be documented in this file. Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
+# Unreleased
+- Fixed: use original asset for filefield thumb when thumb conversion is not available.
+
 ## 0.5.14 - 2021-01-14
 - Fixed: pagination in admin when using filtering.
 - Added: Filefield:uploadButtonLabel() to set a custom label for the upload button